requirements.txt
pandas
numpy
flask
scikit-learn
matplotlib
requests
flask-cors
torch
tqdm
load_dotenv
openai